+++ components/Buttons/Buttons.module.css
@@ -26,11 +26,3 @@
 .iconBtn.modify:hover svg path {
   fill: white;
 }
-/* 
-.iconBtn.delete:hover svg path {
-  fill: white;
-} */
-
-/* .iconBtn.modify:hover svg path {
-  fill: white;
-} */
